fix(flips): widen completed-flip money DTOs int→long so >2.1b flips don't break Flip History

+++ src/main/java/com/flipsmart/api/dto/ActiveFlipsResponse.java
@@ -19,6 +19,6 @@ public class ActiveFlipsResponse
 	private int totalItems;
 
 	@SerializedName("total_invested")
-	private int totalInvested;
+	private long totalInvested;
 }
